+++ src/main/java/com/wanted/needswork/repository/EmployerRepository.java
@@ -4,8 +4,9 @@ import com.wanted.needswork.models.Employer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.swing.text.html.Option;
+import java.math.BigInteger;
 
 public interface EmployerRepository extends JpaRepository <Employer,Integer> {
-    Employer findByUserId(Integer userId);
+    Employer findByUserId(BigInteger userId);
 }
